+++ src/pas/plugins/ldap/plonecontrolpanel/controlpanel.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from pas.plugins.ldap.properties import BasePropertiesForm
 from Products.CMFCore.interfaces import ISiteRoot
+from Products.CMFPlone.resources import add_bundle_on_request
 from Products.statusmessages.interfaces import IStatusMessage
 from zope.component import getUtility
 from zope.i18nmessageid import MessageFactory
@@ -14,6 +15,10 @@ def getPortal():
 
 
 class LDAPControlPanel(BasePropertiesForm):
+    def __init__(self, context, request):
+        super(LDAPControlPanel, self).__init__(context, request)
+        add_bundle_on_request(request, "yafowil")
+
     def next(self, request):
         return "%s/plone_ldapcontrolpanel" % self.context.absolute_url()